cleaning + fixing test
