+++ go/pkg/ca/renewal/grpc/BUILD.bazel
@@ -41,6 +41,7 @@ go_test(
         "//go/lib/scrypto/cppki:go_default_library",
         "//go/lib/scrypto/signed:go_default_library",
         "//go/lib/serrors:go_default_library",
+        "//go/lib/snet:go_default_library",
         "//go/lib/xtest:go_default_library",
         "//go/pkg/ca/api:go_default_library",
         "//go/pkg/ca/renewal:go_default_library",
@@ -51,6 +52,7 @@ go_test(
         "@com_github_stretchr_testify//assert:go_default_library",
         "@com_github_stretchr_testify//require:go_default_library",
         "@org_golang_google_grpc//codes:go_default_library",
+        "@org_golang_google_grpc//peer:go_default_library",
         "@org_golang_google_grpc//status:go_default_library",
     ],
 )
